Basic caching implementation.
Currently only activated for OSM data.

It will break if the location is not "path friendly", so probably
commas, new line, slashes, ... will cause problems. Eventually, we
should store based on hashed values from the server, this should not
be a long term problem.

Only tested on Linux. Hopefully should work on OS X and windows, but I
haven't tested yet. I believe that the features of ifstream and
ofstream are C++ standard. The posix stat and mkdir are used should
work on OSX since it is a UNIX, and I believe also work on windows.

Currenlty caches in "./cache/". This is probably not best, there are
libraries that can point to the right place to cache such as
https://github.com/sago007/PlatformFolders which we should integrate
and use.