[Reddit] default constructor
Removing implementation of default constructor. Specifying default as that should be enough.

There used to be a specific constructor which was taking all the class parameter. But that seems unnecessary.
We could add it for readability purposes, but we do not expect users to ever create their own Reddit objects. So it shouldn't be necessary.

closes #123